Begin adding scripting documentation.
I chose docsify (https://docsify.js.org/) arbitrarily but it seems to look and work pretty well. Please check over the docs for errors since I had to guess on a little bit.

I also chose to write function and property specifications in a typescript-like format for ease of reading.